move out SvLinkNewNodeInput, SvGenericUITooltipOperator, throttle_tree_update
