+++ data_structure.py
@@ -15,6 +15,7 @@
 #  Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 #
 # ##### END GPL LICENSE BLOCK #####
+from contextlib import contextmanager
 from functools import wraps
 from math import radians, ceil
 import itertools
@@ -928,6 +929,28 @@ def update_with_kwargs(update_function, **kwargs):
         handel_update_call.extra_args[attr_name] = data
 
     return handel_update_call
+
+
+@contextmanager
+def throttle_tree_update(node):
+    """ usage
+    from sverchok.node_tree import throttle_tree_update
+
+    inside your node, f.ex inside a wrapped_update that creates a socket
+
+    def wrapped_update(self, context):
+        with throttle_tree_update(self):
+            self.inputs.new(...)
+            self.outputs.new(...)
+
+    that's it.
+
+    """
+    try:
+        node.id_data.skip_tree_update = True
+        yield node
+    finally:
+        node.id_data.skip_tree_update = False
 
 
 ##############################################################
